apps/api/requirements/base.txt
# base requirements

# django
Django==4.2.27
# rest framework
djangorestframework==3.15.2
# postgres
psycopg==3.3.0
psycopg-binary==3.3.0
psycopg-c==3.3.0
dj-database-url==2.1.0
psycopg-pool==3.2.6
# mongo
pymongo==4.6.3
# redis
redis==5.0.4
django-redis==5.4.0
# cors
django-cors-headers==4.3.1
# celery
celery==5.4.0
django_celery_beat==2.6.0
django-celery-results==2.5.1
# file serve
whitenoise==6.11.0
# fake data
faker==25.0.0
# filters
django-filter==24.2
# json model
jsonmodels==2.7.0
# storage
django-storages==1.14.2
# user management
django-crum==0.7.9
# web server
uvicorn==0.29.0
# sockets
channels==4.1.0
# ai
openai==1.63.2
# slack
slack-sdk==3.27.1
# xlsx generation
openpyxl==3.1.2
# logging
python-json-logger==3.3.0
# html parser
beautifulsoup4==4.12.3
# analytics
posthog==3.5.0
# crypto
cryptography==44.0.1
pycryptodome==3.22.0
# html validator
lxml==6.0.0
xmlsec==1.3.16
# s3
boto3==1.34.96
# password validator
zxcvbn==4.4.28
# timezone
pytz==2024.1
# jwt
PyJWT==2.8.0
# OpenTelemetry
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-django==0.49b1
opentelemetry-exporter-otlp==1.28.1
# SAML
python3-saml==1.16.0
# feature flag
openfeature-sdk==0.7.0
# graphql
strawberry-graphql-django==0.48.0
# django rest framework jwt
djangorestframework-simplejwt==5.5.1
# firebase admin
firebase-admin==6.6.0
# oauth toolkit
django-oauth-toolkit==3.0.1

# types
pydantic==2.10.6

# rate limiting
django-ratelimit==4.1.0

# OpenAPI Specification
drf-spectacular==0.28.0

# html sanitizer
nh3==0.2.18

# OpenSearch Integration
opensearch-py==2.8.0
django-opensearch-dsl==0.7.0

# pika for rabbitmq
pika==1.3.2

# psutil for monitoring
psutil==7.0.0

# pgtriggers
django-pgtrigger==4.15.3

# Ldap
python-ldap==3.4.5
